requirements.txt
asgiref==3.3.1
Django==3.0.3
django-crispy-forms==1.10.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-taggit==1.3.0
djangorestframework==3.12.2
importlib-metadata==3.1.0
Markdown==3.3.3
mysqlclient==2.0.1
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
zipp==3.4.0
